refactor: move page builder blocks
