+++ client-app/app-runner.ts
@@ -10,7 +10,7 @@ import { createI18n } from "@/i18n";
 import { createRouter } from "@/router";
 import { getBaseUrl } from "@/core/utilities";
 import App from "./App.vue";
-import PageBuilderBlocks from "@/pages/blocks";
+import { templateBlocks } from "@/shared/static-content";
 import ProductBlocks from "@/shared/catalog/components/product";
 import * as UIKitComponents from "@/ui-kit/components";
 import client from "@/xapi/graphql-client";
@@ -97,7 +97,8 @@ export default async (getPlugins: (options: any) => { plugin: Plugin; options: a
   Object.entries(UIKitComponents).forEach(([name, component]) => app.component(name, component));
 
   // Register Page builder components globally
-  Object.entries(PageBuilderBlocks).forEach(([name, component]) => app.component(name, component));
+  Object.entries(templateBlocks).forEach(([name, component]) => app.component(name, component));
+
   // Register Page builder product components globally
   Object.entries(ProductBlocks).forEach(([name, component]) => app.component(name, component));
