+++ client-app/core/composables/index.ts
@@ -8,5 +8,3 @@ export { default as usePageHead } from "./usePageHead";
 export { default as useRouteQueryParam } from "./useRouteQueryParam";
 export { default as useThemeContext } from "./useThemeContext";
 export { default as useDomUtils } from "./useDomUtils";
-export { default as useStaticPage } from "./useStaticPage";
-export { default as useTemplate } from "./useTemplate";
