+++ client-app/shared/static-content/components/dark-promo-text.vue
