+++ client-app/shared/static-content/components/image.vue
