+++ client-app/shared/static-content/components/index.ts
