+++ client-app/shared/static-content/components/products-block.vue
