+++ client-app/shared/static-content/components/promo-text.vue
