+++ client-app/shared/static-content/components/related-products.vue
