+++ client-app/shared/static-content/components/side-image.vue
