+++ client-app/shared/static-content/components/text.vue
