+++ client-app/shared/static-content/composables/index.ts
@@ -1,0 +1,2 @@
+export { default as useStaticPage } from "./useStaticPage";
+export { default as useTemplate } from "./useTemplate";
