+++ client-app/shared/static-content/index.ts
@@ -1,0 +1,3 @@
+export { default as templateBlocks } from "./components";
+export * from "./composables";
+export * from "./types";
