+++ client-app/shared/static-content/types/index.ts
